Bundle CSS separately
Introduce a plugin, MiniCssExtractPlugin, so that CSS will be bundled separately.

Note that the plugin specified in the video is now deprecated and that this is the successor.

Setup changed a little bit, but not too much.

Docs also spot on for this simple use-case: https://github.com/webpack-contrib/mini-css-extract-plugin#getting-started